+++ TinkoffASDKUI/TinkoffASDKUI/AcquiringUISDK.swift
@@ -59,6 +59,7 @@ public class AcquiringViewConfiguration {
 
     ///
     /// Локализация формы оплаты
+    @available(*, deprecated, message: "Will be removed soon")
     public struct LocalizableInfo {
         var table: String?
         var bundle: Bundle?
@@ -74,6 +75,7 @@ public class AcquiringViewConfiguration {
         }
     }
 
+    @available(*, deprecated, message: "Will be removed soon")
     public var localizableInfo: LocalizableInfo?
 
     ///  Сканер
@@ -186,7 +188,6 @@ public class AcquiringUISDK: NSObject {
                 style: Style = DefaultStyle()) throws {
         acquiringSdk = try AcquiringSdk(configuration: configuration)
         self.style = style
-        AcqLoc.instance.setup(lang: nil, table: nil, bundle: nil)
         self.sbpAssembly = SBPAssembly(coreSDK: acquiringSdk, style: style)
         self.tinkoffPayAssembly = TinkoffPayAssembly(coreSDK: acquiringSdk,
                                                      tinkoffPayStatusCacheLifeTime: configuration.tinkoffPayStatusCacheLifeTime)
@@ -235,8 +236,6 @@ public class AcquiringUISDK: NSObject {
     }
 
     public func presentAddCardView(on presentingViewController: UIViewController, customerKey: String, configuration: AcquiringViewConfiguration, completeHandler: @escaping (_ result: Result<PaymentCard?, Error>) -> Void) {
-        AcqLoc.instance.setup(lang: configuration.localizableInfo?.lang, table: configuration.localizableInfo?.table, bundle: configuration.localizableInfo?.bundle)
-
         self.presentingViewController = presentingViewController
         acquiringViewConfiguration = configuration
 
@@ -477,7 +476,7 @@ public class AcquiringUISDK: NSObject {
         }
 
         presentAcquiringPaymentView(presentingViewController: presentingViewController, customerKey: nil, configuration: configuration) { view in
-            let viewTitle = AcqLoc.instance.localize("TinkoffAcquiring.view.title.payQRCode")
+            let viewTitle = Loc.TinkoffAcquiring.View.Title.payQRCode
             view.changedStatus(.initWaiting)
             self.getStaticQRCode { [weak view] response in
                 switch response {
@@ -488,7 +487,7 @@ public class AcquiringUISDK: NSObject {
 
                 case let .failure(error):
                     DispatchQueue.main.async {
-                        let alertTitle = AcqLoc.instance.localize("TinkoffAcquiring.alert.title.error")
+                        let alertTitle = Loc.TinkoffAcquiring.Alert.Title.error
 
                         if let alert = configuration.alertViewHelper?.presentAlertView(alertTitle, message: error.localizedDescription, dismissCompletion: nil) {
                             view?.closeVC(animated: true) {
@@ -692,7 +691,7 @@ public class AcquiringUISDK: NSObject {
     private func presentApplePayActivity(_ request: PKPaymentRequest) {
         guard let viewController = PKPaymentAuthorizationViewController(paymentRequest: request) else {
             acquiringView?.closeVC(animated: true) {
-                let error = NSError(domain: AcqLoc.instance.localize("TinkoffAcquiring.unknown.response.status"),
+                let error = NSError(domain: Loc.TinkoffAcquiring.Unknown.Response.status,
                                     code: 0,
                                     userInfo: nil)
 
@@ -725,7 +724,7 @@ public class AcquiringUISDK: NSObject {
                         if UIApplication.shared.canOpenURL(url) {
                             UIApplication.shared.open(url, options: [:]) { _ in
                                 self?.sbpWaitingIncominPayment(paymentId: paymentId, source: qrCodeResponse.qrCodeData, sourceType: paymentInvoiceSource)
-                                self?.acquiringView?.changedStatus(.paymentWaitingSBPUrl(url: qrCodeResponse.qrCodeData, status: "Выбор источника оплаты"))
+                                self?.acquiringView?.changedStatus(.paymentWaitingSBPUrl(url: qrCodeResponse.qrCodeData, status: Loc.TinkoffAcquiring.Text.Status.selectingPaymentSource))
                             }
                         } else {
                             let activityViewController = UIActivityViewController(activityItems: [url], applicationActivities: [])
@@ -736,7 +735,7 @@ public class AcquiringUISDK: NSObject {
                             }
 
                             self?.acquiringView?.presentVC(activityViewController, animated: true, completion: {
-                                self?.acquiringView?.changedStatus(.paymentWaitingSBPUrl(url: qrCodeResponse.qrCodeData, status: "Выбор источника оплаты"))
+                                self?.acquiringView?.changedStatus(.paymentWaitingSBPUrl(url: qrCodeResponse.qrCodeData, status: Loc.TinkoffAcquiring.Text.Status.selectingPaymentSource))
                             })
                         }
                     } else {
@@ -749,7 +748,7 @@ public class AcquiringUISDK: NSObject {
                 DispatchQueue.main.async {
                     self?.acquiringView?.changedStatus(.error(error))
 
-                    let alertTitle = AcqLoc.instance.localize("TinkoffAcquiring.alert.title.error")
+                    let alertTitle = Loc.TinkoffAcquiring.Alert.Title.error
                     if let alert = configuration.alertViewHelper?.presentAlertView(alertTitle, message: error.localizedDescription, dismissCompletion: nil) {
                         self?.presentingViewController?.present(alert, animated: true, completion: {
                             //
@@ -784,9 +783,9 @@ public class AcquiringUISDK: NSObject {
         }
 
         if sourceType == .url {
-            acquiringView?.changedStatus(.paymentWaitingSBPUrl(url: source, status: "Ожидание оплаты"))
+            acquiringView?.changedStatus(.paymentWaitingSBPUrl(url: source, status: Loc.TinkoffAcquiring.Text.Status.waitingPayment))
         } else {
-            acquiringView?.changedStatus(.paymentWaitingSBPQrCode(qrCode: source, status: "Ожидание оплаты"))
+            acquiringView?.changedStatus(.paymentWaitingSBPQrCode(qrCode: source, status: Loc.TinkoffAcquiring.Text.Status.waitingPayment))
         }
 
         checkPaymentStatus?.fetchStatus(completionStatus: completionStatus)
@@ -804,7 +803,6 @@ public class AcquiringUISDK: NSObject {
         onPresenting: (((AcquiringView) -> Void))? = nil
     ) {
         self.presentingViewController = presentingViewController
-        AcqLoc.instance.setup(lang: configuration.localizableInfo?.lang, table: configuration.localizableInfo?.table, bundle: configuration.localizableInfo?.bundle)
 
         // create
         let modalViewController = AcquiringPaymentViewController(nibName: "AcquiringPaymentViewController",
@@ -1211,7 +1209,7 @@ public class AcquiringUISDK: NSObject {
                     completionHandler(.success(response))
 
                 case .unknown:
-                    let error = NSError(domain: finishResult.errorMessage ?? AcqLoc.instance.localize("TinkoffAcquiring.unknown.response.status"),
+                    let error = NSError(domain: finishResult.errorMessage ?? Loc.TinkoffAcquiring.Unknown.Response.status,
                                         code: finishResult.errorCode,
                                         userInfo: nil)
 
@@ -1324,7 +1322,7 @@ public class AcquiringUISDK: NSObject {
         } else {
             let paymentCanceledResponse = PaymentStatusResponse(success: false,
                                                                 errorCode: 0,
-                                                                errorMessage: AcqLoc.instance.localize("TinkoffAcquiring.alert.message.addingCardCancel"),
+                                                                errorMessage: Loc.TinkoffAcquiring.Alert.Message.addingCardCancel,
                                                                 orderId: "",
                                                                 paymentId: 0,
                                                                 amount: 0,
@@ -1399,7 +1397,7 @@ public class AcquiringUISDK: NSObject {
                         })
                     case let .failure(error):
                         viewController?.dismiss(animated: true, completion: {
-                            let alertTitle = AcqLoc.instance.localize("TinkoffAcquiring.alert.title.error")
+                            let alertTitle = Loc.TinkoffAcquiring.Alert.Title.error
                             if let alert = alertViewHelper?.presentAlertView(alertTitle, message: error.localizedDescription, dismissCompletion: nil) {
                                 self?.presentingViewController?.presentOnTop(viewController: alert, animated: true)
                             } else {
@@ -1501,7 +1499,6 @@ extension AcquiringUISDK: AcquiringCardListDataSourceDelegate {
                         customerKey: String,
                         configuration: AcquiringViewConfiguration,
                         completeHandler: @escaping (Result<PaymentCard?, Error>) -> Void) {
-        AcqLoc.instance.setup(lang: configuration.localizableInfo?.lang, table: configuration.localizableInfo?.table, bundle: configuration.localizableInfo?.bundle)
 
         self.presentingViewController = presentingViewController
         acquiringViewConfiguration = configuration
@@ -1624,7 +1621,7 @@ extension AcquiringUISDK: AcquiringCardListDataSourceDelegate {
             confirmationComplete(.success(response))
 
         case .unknown:
-            let error = NSError(domain: confirmationResponse.errorMessage ?? AcqLoc.instance.localize("TinkoffAcquiring.unknown.response.status"),
+            let error = NSError(domain: confirmationResponse.errorMessage ?? Loc.TinkoffAcquiring.Unknown.Response.status,
                                 code: confirmationResponse.errorCode, userInfo: nil)
 
             confirmationComplete(.failure(error))
@@ -1667,11 +1664,6 @@ extension AcquiringUISDK: AcquiringCardListDataSourceDelegate {
         customerKey: String,
         configuration: AcquiringViewConfiguration
     ) {
-        AcqLoc.instance.setup(
-            lang: configuration.localizableInfo?.lang,
-            table: configuration.localizableInfo?.table,
-            bundle: configuration.localizableInfo?.bundle
-        )
 
         if acquiringViewConfiguration == nil {
             acquiringViewConfiguration = configuration
@@ -1779,7 +1771,7 @@ extension AcquiringUISDK: WKNavigationDelegate {
                     // decode as a default `AcquiringResponse`
                     guard let acquiringResponse = try? JSONDecoder().decode(AcquiringResponse.self, from: data) else {
                         self?.webViewController?.dismiss(animated: true, completion: { [weak self] in
-                            let error = NSError(domain: AcqLoc.instance.localize("TinkoffAcquiring.unknown.response.status"), code: 0, userInfo: nil)
+                            let error = NSError(domain: Loc.TinkoffAcquiring.Unknown.Response.status, code: 0, userInfo: nil)
                             self?.on3DSCheckingCompletionHandler?(.failure(error))
                             self?.on3DSCheckingAddCardCompletionHandler?(.failure(error))
                         })
@@ -1789,7 +1781,7 @@ extension AcquiringUISDK: WKNavigationDelegate {
 
                     // data  in `AcquiringResponse` format but `Success = 0;` ( `false` )
                     guard acquiringResponse.success else {
-                        let error = NSError(domain: acquiringResponse.errorMessage ?? AcqLoc.instance.localize("TinkoffAcquiring.response.success.false"),
+                        let error = NSError(domain: acquiringResponse.errorMessage ?? Loc.TinkoffAcquiring.Unknown.Error.status,
                                             code: acquiringResponse.errorCode,
                                             userInfo: try? acquiringResponse.encode2JSONObject())
 
@@ -1804,7 +1796,7 @@ extension AcquiringUISDK: WKNavigationDelegate {
                     // data in `PaymentStatusResponse` format
                     if self?.on3DSCheckingCompletionHandler != nil {
                         guard let responseObject: PaymentStatusResponse = try? JSONDecoder().decode(PaymentStatusResponse.self, from: data) else {
-                            let error = NSError(domain: acquiringResponse.errorMessage ?? AcqLoc.instance.localize("TinkoffAcquiring.response.success.false"),
+                            let error = NSError(domain: acquiringResponse.errorMessage ?? Loc.TinkoffAcquiring.Unknown.Error.status,
                                                 code: acquiringResponse.errorCode,
                                                 userInfo: try? acquiringResponse.encode2JSONObject())
 
@@ -1824,7 +1816,7 @@ extension AcquiringUISDK: WKNavigationDelegate {
                     // data in `AddCardStatusResponse` format
                     if self?.on3DSCheckingAddCardCompletionHandler != nil {
                         guard let responseObject: AddCardStatusResponse = try? JSONDecoder().decode(AddCardStatusResponse.self, from: data) else {
-                            let error = NSError(domain: acquiringResponse.errorMessage ?? AcqLoc.instance.localize("TinkoffAcquiring.response.success.false"),
+                            let error = NSError(domain: acquiringResponse.errorMessage ?? Loc.TinkoffAcquiring.Unknown.Error.status,
                                                 code: acquiringResponse.errorCode,
                                                 userInfo: try? acquiringResponse.encode2JSONObject())